comment for benchmark
# Conflicts:
#	services/gitdiff/highlightdiff_test.go
